# shared/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: ChatMod Mobile API
  version: 0.1.0
  description: Backend contract for the phone-hosted YouTube Live moderation bot.
servers:
  - url: http://localhost:4100
paths:
...
  /youtube/connect-url:
    get:
      summary: Get the OAuth connection URL
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OAuth URL payload
  /youtube/account:
    get:
      summary: Get connected YouTube account and channel metadata
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Connected YouTube account status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/YouTubeAccountStatus"
  /youtube/oauth/callback:
    get:
      summary: Exchange YouTube OAuth callback code
      responses:
        "200":
          description: YouTube connection result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/YouTubeConnectionResult"
        "503":
          description: YouTube OAuth is not configured
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    adminApiKey:
      type: apiKey
      in: header
      name: x-admin-api-key
  schemas:
...
    YouTubeConnectionResult:
      type: object
      required:
        - connected
        - deviceId
        - tokensStored
        - linkedAccountId
        - channelId
        - channelTitle
        - tokenExpiresAt
        - scopes
      properties:
        connected:
          type: boolean
        deviceId:
          type: string
        tokensStored:
          type: boolean
        linkedAccountId:
          type:
            - string
            - "null"
        channelId:
          type:
            - string
            - "null"
        channelTitle:
          type:
            - string
            - "null"
        tokenExpiresAt:
          type:
            - string
            - "null"
          format: date-time
        scopes:
          type: array
          items:
            type: string
    YouTubeAccountStatus:
      type: object
      required:
        - configured
        - source
        - account
      properties:
        configured:
          type: boolean
        source:
          type: string
          enum:
            - mock
            - youtube-api
        account:
          $ref: "#/components/schemas/YouTubeLinkedAccountStatus"
    YouTubeLinkedAccountStatus:
      type: object
      required:
        - connected
        - linkedAccountId
        - channelId
        - channelTitle
        - hasAccessToken
        - hasRefreshToken
        - tokenExpiresAt
      properties:
        connected:
          type: boolean
        linkedAccountId:
          type:
            - string
            - "null"
        channelId:
          type:
            - string
            - "null"
        channelTitle:
          type:
            - string
            - "null"
        hasAccessToken:
          type: boolean
        hasRefreshToken:
          type: boolean
        tokenExpiresAt:
          type:
            - string
            - "null"
          format: date-time
...